requirements.txt
selenium==4.22.0
beautifulsoup4==4.12.2
pandas==2.1.4
mysql-connector-python==9.0.0
python-dotenv==0.21.0